rfq_system/requirements.txt
Django==5.0.7
reportlab==4.2.2
python-dotenv==1.0.1